refactor: move scope module to src/ from src/oauth2
