+++ examples/sso.rs
@@ -122,7 +122,7 @@ async fn main() {
 
 async fn login(session: Session, Extension(esi_client): Extension<eve_esi::Client>) -> Response {
     // Build the scopes we wish to request from the user
-    let scopes = eve_esi::oauth2::ScopeBuilder::new().public_data().build();
+    let scopes = eve_esi::ScopeBuilder::new().public_data().build();
 
     // Generate the login url or return an error if one occurs
     let login_url = match esi_client.oauth2().login_url(scopes) {
