+++ src/lib.rs
@@ -104,13 +104,14 @@ pub mod error;
 pub mod esi;
 pub mod model;
 pub mod oauth2;
+pub mod scope;
 
 pub use crate::builder::ClientBuilder;
 pub use crate::client::Client;
 pub use crate::config::{Config, ConfigBuilder};
 pub use crate::error::{ConfigError, Error};
 pub use crate::oauth2::error::OAuthError;
-pub use crate::oauth2::ScopeBuilder;
+pub use crate::scope::ScopeBuilder;
 
 mod constant;
