+++ tests/endpoints/corporation.rs
@@ -1,4 +1,4 @@
-use eve_esi::oauth2::scope::{CorporationScopes, WalletScopes};
+use eve_esi::scope::{CorporationScopes, WalletScopes};
 use eve_esi::ScopeBuilder;
 
 use crate::endpoints::util::{authenticated_endpoint_test_setup, mock_access_token_with_scopes};
